feat: more adjustments
